Move TestDetektion to detekt/test package
