+++ detekt-cli/src/test/kotlin/io/gitlab/arturbosch/detekt/cli/console/NotificationReportSpec.kt
@@ -1,6 +1,6 @@
 package io.gitlab.arturbosch.detekt.cli.console
 
-import io.gitlab.arturbosch.detekt.cli.TestDetektion
+import io.gitlab.arturbosch.detekt.test.TestDetektion
 import io.gitlab.arturbosch.detekt.cli.createNotification
 import io.gitlab.arturbosch.detekt.test.resource
 import org.assertj.core.api.Assertions.assertThat
